requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.5.1
python-jose==3.3.0
passlib==1.7.4
pydantic==2.8.2
bcrypt==4.0.1
